Add feature to display single country details without navigating to a details page
