+++ index.html
@@ -9,7 +9,7 @@
     />
     <title>Countries React</title>
   </head>
-  <body>
+  <body >
     <div id="root"></div>
     <script type="module" src="src/main.jsx"></script>
   </body>
