+++ src/App.jsx
@@ -2,18 +2,14 @@ import "../src/assets/css/common.css";
 import "../src/assets/css/details.css";
 import "../src/assets/css/main.css";
 import "../src/assets/scss/common.scss";
-import Home from "./pages/Home";
-import Header from "./components/Header.jsx";
+import Home from "./Home.jsx";
+
 
 function App() {
     return (
-
         <div>
-
             <Home/>
         </div>
-
-
     );
 }
