+++ src/assets/css/common.css
@@ -7,19 +7,19 @@
   outline: none;
 }
 
-html {
-  font-size: 100%;
-  scroll-behavior: smooth;
-}
+/*html {*/
+/*  font-size: 100%;*/
+/*  scroll-behavior: smooth;*/
+/*}*/
 
-body {
-  font-family: 'Nunito Sans', sans-serif;
-  font-weight: 600;
-  font-size: 14px;
-  background-color: #fafafa;
-  color: #111517;
-  transition: 0.2s linear;
-}
+/*body {*/
+/*  font-family: 'arial', sans-serif;*/
+/*  font-weight: 600;*/
+/*  font-size: 14px;*/
+/*  background-color: #fafafa;*/
+/*  color: #111517;*/
+/*  transition: 0.2s linear;*/
+/*}*/
 
 a {
   text-decoration: none;
