+++ src/assets/css/details.css
@@ -1,5 +1,6 @@
 body {
   font-size: 16px;
+
 }
 
 .btn:hover {
